requirements.txt
requests==2.31.0
schedule==1.2.1
gspread==6.0.2
google-auth==2.28.0
anthropic==0.21.3
openai==1.14.0
python-dotenv==1.0.1
feedparser==6.0.11
